refactor: lift expansion state, extract tree builder, and fix effect dependencies
